pyproject.toml
[project]
name = "fireflyiii-enricher-core"
version = "0.2.0"
description = "A core library for enriching Firefly III transactions"
authors = [{ name = "WiNi83", email = "rico.gotti@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

dependencies = [
    "requests",
    "python-dotenv"
]

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
